moves the configuration files to shared/configuration folder
