+++ src/app.js
@@ -35,7 +35,6 @@ import '../public/images/icon-lantaarn@3x.png';
 import 'leaflet/dist/leaflet.css';
 import 'leaflet.markercluster/dist/MarkerCluster.css';
 import 'leaflet.markercluster/dist/MarkerCluster.Default.css';
-import 'static/nlmaps.css';
 import 'amsterdam-stijl/dist/css/ams-stijl.css';
 import history from 'utils/history';
