+++ src/shared/configuration/categories.js
